requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.1
langchain>=0.1.0
langchain-core>=0.1.10
langchain-google-genai>=0.0.11
google-generativeai>=0.4.1
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
pypdf>=3.17.1
python-docx>=1.1.0
